+++ geotrek/feedback/filters.py
@@ -1,5 +1,5 @@
 from django.utils.translation import gettext_lazy as _
-from django_filters import MultipleChoiceFilter
+from django_filters import ChoiceFilter, MultipleChoiceFilter
 from mapentity.filters import MapEntityFilterSet
 
 from geotrek.zoning.filters import ZoningFilterSet
@@ -18,14 +18,20 @@ class ReportFilterSet(ZoningFilterSet, MapEntityFilterSet):
         field_name='date_update__year',
         choices=lambda: Report.objects.year_update_choices()
     )
+    provider = ChoiceFilter(
+        field_name='provider',
+        empty_label=_("Provider"),
+        label=_("Provider"),
+        choices=lambda: Report.objects.provider_choices()
+    )
 
     class Meta(MapEntityFilterSet.Meta):
         model = Report
-        fields = ['activity', 'category', 'status', 'problem_magnitude', 'assigned_user']
+        fields = ['activity', 'category', 'status', 'problem_magnitude', 'assigned_user', 'provider']
 
 
 class ReportEmailFilterSet(ReportFilterSet):
 
     class Meta(MapEntityFilterSet.Meta):
         model = Report
-        fields = ['activity', 'email', 'category', 'status', 'problem_magnitude', 'assigned_user']
+        fields = ['activity', 'email', 'category', 'status', 'problem_magnitude', 'assigned_user', 'provider']