+++ geotrek/feedback/forms.py
@@ -165,5 +165,7 @@ class ReportForm(CommonForm):
                 report.change_position_in_suricate(force=force_gps)
         elif report.status and report.uses_timers and (creation or self.old_status != report.status):  # Outside of workflow, create timer if report is new or if its status changed
             TimerEvent.objects.create(step=report.status, report=report)
-
+        elif creation and not settings.SURICATE_WORKFLOW_ENABLED:
+            report.provider = "Geotrek-admin"
+            report.save()
         return report